+++ src/contexts/Balances/index.tsx
@@ -17,6 +17,8 @@ import { SyncController } from 'controllers/Sync';
 import { useApi } from 'contexts/Api';
 import { ActivePoolsController } from 'controllers/ActivePools';
 import { useCreatePoolAccounts } from 'hooks/useCreatePoolAccounts';
+import { useNetwork } from 'contexts/Network';
+import { SubscriptionsController } from 'controllers/Subscriptions';
 
 export const BalancesContext = createContext<BalancesContextInterface>(
   defaults.defaultBalancesContext
@@ -25,6 +27,7 @@ export const BalancesContext = createContext<BalancesContextInterface>(
 export const useBalances = () => useContext(BalancesContext);
 
 export const BalancesProvider = ({ children }: { children: ReactNode }) => {
+  const { network } = useNetwork();
   const { getBondedAccount } = useBonded();
   const { accounts } = useImportedAccounts();
   const createPoolAccounts = useCreatePoolAccounts();
@@ -82,7 +85,8 @@ export const BalancesProvider = ({ children }: { children: ReactNode }) => {
 
   // Check whether all accounts have been synced and update state accordingly.
   const checkBalancesSynced = () => {
-    if (Object.keys(BalancesController.balances).length === accounts.length) {
+    // Check if subscriptions have been assigned to all accounts.
+    if (BalancesController.areAccountsSynced(network)) {
       SyncController.dispatch('balances', 'complete');
     }
   };
@@ -91,7 +95,11 @@ export const BalancesProvider = ({ children }: { children: ReactNode }) => {
   // payload.
   const getNonce = (address: MaybeAddress) => {
     if (address) {
-      const maybeNonce = BalancesController.balances[address]?.nonce;
+      const maybeNonce = SubscriptionsController.get(
+        network,
+        `balances-${address}`
+      )?.balances[address]?.nonce;
+
       if (maybeNonce) {
         return maybeNonce;
       }
