+++ src/contexts/Bonded/defaults.ts
@@ -2,8 +2,8 @@
 // SPDX-License-Identifier: GPL-3.0-only
 /* eslint-disable @typescript-eslint/no-unused-vars */
 
-import type { Nominations } from 'contexts/Balances/types';
 import type { BondedContextInterface } from 'contexts/Bonded/types';
+import type { Nominations } from 'model/Subscribe/Balance/types';
 
 export const nominations: Nominations = {
   targets: [],
