+++ src/contexts/Pools/ActivePool/types.ts
@@ -5,7 +5,7 @@ import type BigNumber from 'bignumber.js';
 import type { PoolAddresses } from '../BondedPools/types';
 import type { MaybeAddress } from '@w3ux/react-connect-kit/types';
 import type { Identity, SuperIdentity } from 'contexts/Validators/types';
-import type { Nominations } from 'contexts/Balances/types';
+import type { Nominations } from 'model/Subscribe/Balance/types';
 
 export interface ActivePoolContextState {
   isBonding: () => boolean;
