+++ src/controllers/Balances/index.ts
@@ -1,23 +1,11 @@
 // Copyright 2024 @polkadot-cloud/polkadot-staking-dashboard authors & contributors
 // SPDX-License-Identifier: GPL-3.0-only
 
-import { rmCommas } from '@w3ux/utils';
-import BigNumber from 'bignumber.js';
-import type { AnyApi, MaybeAddress } from 'types';
-import type {
-  ActiveBalance,
-  Balances,
-  Ledger,
-  Nominations,
-  UnlockChunkRaw,
-} from 'contexts/Balances/types';
-import type { PayeeConfig, PayeeOptions } from 'contexts/Setup/types';
-import type { PoolMembership } from 'contexts/Pools/types';
+import type { NetworkName } from 'types';
+import type { ActiveBalance } from 'contexts/Balances/types';
 import { SyncController } from 'controllers/Sync';
-import { defaultNominations } from './defaults';
-import type { VoidFn } from '@polkadot/api/types';
-import type { ApiPromise } from '@polkadot/api';
-import { stringToBn } from 'library/Utils';
+import { Balance } from 'model/Subscribe/Balance';
+import { SubscriptionsController } from 'controllers/Subscriptions';
 
 export class BalancesController {
   // ------------------------------------------------------
@@ -27,35 +15,17 @@ export class BalancesController {
   // Accounts that are being subscribed to.
   static accounts: string[] = [];
 
-  // Account ledgers, populated by api callbacks.
-  static ledgers: Record<string, Ledger> = {};
-
-  // Account balances, populated by api callbacks.
-  static balances: Record<string, Balances> = {};
-
-  // Account payees, populated by api callbacks.
-  static payees: Record<string, PayeeConfig> = {};
-
-  // Account pool membership and claim commissions, populated by api callbacks.
-  static poolMemberships: Record<string, PoolMembership> = {};
-
-  // Account nominations, populated by api callbacks.
-  static nominations: Record<string, Nominations> = {};
-
-  // Unsubscribe objects.
-  static #unsubs: Record<string, VoidFn> = {};
-
   // ------------------------------------------------------
   // Account syncing.
   // ------------------------------------------------------
 
   // Subscribes new accounts and unsubscribes & removes removed accounts.
   static syncAccounts = async (
-    api: ApiPromise,
+    network: NetworkName,
     newAccounts: string[]
   ): Promise<void> => {
     // Handle accounts that have been removed.
-    this.handleRemovedAccounts(newAccounts);
+    this.handleRemovedAccounts(network, newAccounts);
 
     // Determine new accounts that need to be subscribed to.
     const accountsAdded = newAccounts.filter(
@@ -74,241 +44,56 @@ export class BalancesController {
     accountsAdded.forEach(async (address) => {
       this.accounts.push(address);
 
-      // TODO: Move to a `Subscription`.
-      // Move subscription logic to `Subscribe` class.
-      // Listen to subscription changes via event listener in this class.
-
-      const unsub = await api.queryMulti(
-        [
-          [api.query.staking.ledger, address],
-          [api.query.system.account, address],
-          [api.query.balances.locks, address],
-          [api.query.staking.payee, address],
-          [api.query.nominationPools.poolMembers, address],
-          [api.query.nominationPools.claimPermissions, address],
-          [api.query.staking.nominators, address],
-        ],
-        async ([
-          ledgerResult,
-          accountResult,
-          locksResult,
-          payeeResult,
-          poolMembersResult,
-          claimPermissionsResult,
-          nominatorsResult,
-        ]): Promise<void> => {
-          this.handleLedgerCallback(address, ledgerResult);
-          this.handleAccountCallback(address, accountResult, locksResult);
-          this.handlePayeeCallback(address, payeeResult);
-
-          // NOTE: async: contains runtime call for pending rewards.
-          await this.handlePoolMembershipCallback(
-            api,
-            address,
-            poolMembersResult,
-            claimPermissionsResult
-          );
-
-          this.handleNominations(address, nominatorsResult);
-
-          // Send updated account state back to UI.
-          document.dispatchEvent(
-            new CustomEvent('new-account-balance', {
-              detail: {
-                address,
-                ledger: this.ledgers[address],
-                balances: this.balances[address],
-                payee: this.payees[address],
-                poolMembership: this.poolMemberships[address],
-                nominations: this.nominations[address],
-              },
-            })
-          );
-        }
+      // Initialise balance subscription.
+      SubscriptionsController.set(
+        network,
+        `balances-${address}`,
+        new Balance(network, address)
       );
-      this.#unsubs[address] = unsub;
     });
   };
 
   // Remove accounts that no longer exist.
-  static handleRemovedAccounts = (newAccounts: string[]): void => {
+  static handleRemovedAccounts = (
+    network: NetworkName,
+    newAccounts: string[]
+  ): void => {
     // Determine removed accounts.
     const accountsRemoved = this.accounts.filter(
       (account) => !newAccounts.includes(account)
     );
     // Unsubscribe from removed account subscriptions.
     accountsRemoved.forEach((account) => {
-      if (this.#unsubs[account]) {
-        this.#unsubs[account]();
-      }
-      delete this.#unsubs[account];
-      delete this.ledgers[account];
-      delete this.balances[account];
-      delete this.payees[account];
-      delete this.poolMemberships[account];
-      delete this.nominations[account];
+      SubscriptionsController.remove(network, `balances-${account}`);
     });
+
     // Remove removed accounts from class.
     this.accounts = this.accounts.filter(
       (account) => !accountsRemoved.includes(account)
     );
   };
 
-  // Handle ledger callback.
-  static handleLedgerCallback = (address: string, result: AnyApi): void => {
-    const ledger = result.unwrapOr(null);
-
-    // If ledger is null, remove from class data and exit early.
-    if (ledger === null) {
-      delete this.ledgers[address];
-      return;
-    }
-
-    const { stash, total, active, unlocking } = ledger;
-
-    // Send stash address to UI as event if not presently imported.
-    if (!this.accounts.includes(stash.toString())) {
-      document.dispatchEvent(
-        new CustomEvent('new-external-account', {
-          detail: { address: stash.toString() },
-        })
-      );
-    }
-
-    this.ledgers[address] = {
-      stash: stash.toString(),
-      active: stringToBn(active.toString()),
-      total: stringToBn(total.toString()),
-      unlocking: unlocking.toHuman().map(({ era, value }: UnlockChunkRaw) => ({
-        era: Number(rmCommas(era)),
-        value: stringToBn(value),
-      })),
-    };
-  };
-
-  // Handle account callback.
-  static handleAccountCallback = (
-    address: string,
-    { data: accountData, nonce }: AnyApi,
-    locksResult: AnyApi
-  ): void => {
-    this.balances[address] = {
-      nonce: nonce.toNumber(),
-      balance: {
-        free: stringToBn(accountData.free.toString()),
-        reserved: stringToBn(accountData.reserved.toString()),
-        frozen: stringToBn(accountData.frozen.toString()),
-      },
-      locks: locksResult
-        .toHuman()
-        .map((lock: { id: string; amount: string }) => ({
-          ...lock,
-          id: lock.id.trim(),
-          amount: stringToBn(lock.amount),
-        })),
-    };
-  };
-
-  // Handle payee callback. payee with `Account` type is returned as an key value pair, with all
-  // others strings. This function handles both cases and formats into a unified structure.
-  static handlePayeeCallback = (address: string, result: AnyApi): void => {
-    const payeeHuman = result.toHuman();
-    let payeeFinal: PayeeConfig;
-
-    if (payeeHuman !== null) {
-      if (typeof payeeHuman === 'string') {
-        const destination = payeeHuman as PayeeOptions;
-        payeeFinal = {
-          destination,
-          account: null,
-        };
-      } else {
-        const payeeEntry = Object.entries(payeeHuman);
-        const destination = `${payeeEntry[0][0]}` as PayeeOptions;
-        const account = `${payeeEntry[0][1]}` as MaybeAddress;
-        payeeFinal = {
-          destination,
-          account,
-        };
-      }
-      this.payees[address] = payeeFinal;
-    }
-  };
-
-  // Handle pool membership and claim commission callback.
-  static handlePoolMembershipCallback = async (
-    api: ApiPromise,
-    address: string,
-    poolMembersResult: AnyApi,
-    claimPermissionsResult: AnyApi
-  ): Promise<void> => {
-    // If pool membership is `null`, remove pool membership data from class data and exit early.
-    // This skips claim permission data as well as user would not have claim permissions if they are
-    // not in a pool.
-    const membership = poolMembersResult?.unwrapOr(undefined)?.toHuman();
-    if (!membership) {
-      delete this.poolMemberships[address];
-      return;
-    }
-
-    // Format pool membership data.
-    const unlocking = Object.entries(membership?.unbondingEras || {}).map(
-      ([e, v]) => ({
-        era: Number(rmCommas(e as string)),
-        value: new BigNumber(rmCommas(v as string)),
-      })
-    );
-    membership.points = rmCommas(membership?.points || '0');
-    const balance = new BigNumber(
-      (
-        await api.call.nominationPoolsApi.pointsToBalance(
-          membership.poolId,
-          membership.points
-        )
-      )?.toString() || '0'
-    );
-    const claimPermission =
-      claimPermissionsResult?.toString() || 'Permissioned';
-
-    // Persist formatted pool membership data to class.
-    this.poolMemberships[address] = {
-      ...membership,
-      address,
-      balance,
-      claimPermission,
-      unlocking,
-    };
-  };
-
-  // Handle nominations callback.
-  static handleNominations = (
-    address: string,
-    nominatorsResult: AnyApi
-  ): void => {
-    const nominators = nominatorsResult.unwrapOr(null);
-
-    this.nominations[address] =
-      nominators === null
-        ? defaultNominations
-        : {
-            targets: nominators.targets.toHuman(),
-            submittedIn: nominators.submittedIn.toHuman(),
-          };
-  };
-
   // Gets an `ActiveBalance` from class members for the given address if it exists.
-  static getAccountBalances = (address: string): ActiveBalance | undefined => {
-    const ledger = this.ledgers[address];
-    const balances = this.balances[address];
-    const payee = this.payees[address];
-    const poolMembership = this.poolMemberships[address];
-    const nominations = this.nominations[address];
+  static getAccountBalances = (
+    network: NetworkName,
+    address: string
+  ): ActiveBalance | undefined => {
+    const subscription = SubscriptionsController.get(
+      network,
+      `balances-${address}`
+    ) as Balance;
 
-    if (balances === undefined) {
-      // Account info has not synced yet. Note that `ledger` may not exist and therefore cannot be
-      // tested.
+    if (subscription === undefined) {
+      // Account info has not synced yet.
       return undefined;
     }
+
+    const ledger = subscription.ledger;
+    const balances = subscription.balances;
+    const payee = subscription.payee;
+    const poolMembership = subscription.poolMembership;
+    const nominations = subscription.nominations;
+
     return {
       ledger,
       balances,
@@ -318,17 +103,15 @@ export class BalancesController {
     };
   };
 
-  // ------------------------------------------------------
-  // Subscription handling.
-  // ------------------------------------------------------
-
-  // Unsubscribe from all subscriptions and reset class members.
-  static unsubscribe = (): void => {
-    Object.values(this.#unsubs).forEach((unsub) => {
-      unsub();
+  // Checks if all account balance subscriptions have initialised.
+  static areAccountsSynced = (network: NetworkName): boolean =>
+    this.accounts.every((address) => {
+      const subscription = SubscriptionsController.get(
+        network,
+        `balances-${address}`
+      ) as Balance;
+      return subscription.balances !== undefined;
     });
-    this.#unsubs = {};
-  };
 
   // ------------------------------------------------------
   // Class helpers.
