+++ src/hooks/useActivePools/types.ts
@@ -1,9 +1,9 @@
 // Copyright 2024 @polkadot-cloud/polkadot-staking-dashboard authors & contributors
 // SPDX-License-Identifier: GPL-3.0-only
 
-import type { Nominations } from 'contexts/Balances/types';
 import type { ActivePool } from 'contexts/Pools/ActivePool/types';
 import type { DetailActivePool } from 'controllers/ActivePools/types';
+import type { Nominations } from 'model/Subscribe/Balance/types';
 import type { MaybeAddress } from 'types';
 
 export interface ActivePoolsProps {
