+++ src/library/Form/types.ts
@@ -2,13 +2,13 @@
 // SPDX-License-Identifier: GPL-3.0-only
 
 import type BigNumber from 'bignumber.js';
-import type { Balance } from 'contexts/Balances/types';
 import type {
   ExtensionAccount,
   ExternalAccount,
 } from '@w3ux/react-connect-kit/types';
 import type { BondFor, MaybeAddress } from 'types';
 import type { ClaimPermission } from 'contexts/Pools/types';
+import type { Balance } from 'model/Subscribe/Balance';
 
 export interface ExtensionAccountItem extends ExtensionAccount {
   active?: boolean;
